requirements.txt
langchain
ipykernel
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-ollama
chromadb
python-dotenv
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit